+++ src/raster/rasterband.rs
@@ -5,11 +5,13 @@ use crate::raster::{GDALDataType, GdalType};
 use crate::utils::{_last_cpl_err, _last_null_pointer_err, _string};
 use gdal_sys::{
     self, CPLErr, GDALColorEntry, GDALColorInterp, GDALColorTableH, GDALComputeRasterMinMax,
+    GDALCreateColorRamp, GDALCreateColorTable, GDALDestroyColorTable, GDALGetPaletteInterpretation,
     GDALGetRasterStatistics, GDALMajorObjectH, GDALPaletteInterp, GDALRIOResampleAlg, GDALRWFlag,
-    GDALRasterBandH, GDALRasterIOExtraArg,
+    GDALRasterBandH, GDALRasterIOExtraArg, GDALSetColorEntry, GDALSetRasterColorTable,
 };
 use libc::c_int;
 use std::ffi::CString;
+use std::fmt::{Debug, Formatter};
 use std::marker::PhantomData;
 
 #[cfg(feature = "ndarray")]
@@ -485,6 +487,13 @@ impl<'a> RasterBand<'a> {
         Some(ColorTable::from_c_color_table(c_color_table))
     }
 
+    /// Set the color table for this band.
+    ///
+    /// See [`ColorTable`] for usage example.
+    pub fn set_color_table(&mut self, colors: &ColorTable) {
+        unsafe { GDALSetRasterColorTable(self.c_rasterband, colors.c_color_table) };
+    }
+
     /// Returns the scale of this band if set.
     pub fn scale(&self) -> Option<f64> {
         let mut pb_success = 1;
@@ -811,15 +820,21 @@ impl ColorInterpretation {
     }
 }
 
+/// Types of color interpretations for a [`ColorTable`].
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum PaletteInterpretation {
+    /// Grayscale
     Gray,
+    /// Red, Green, Blue and Alpha
     Rgba,
+    /// Cyan, Magenta, Yellow and Black
     Cmyk,
+    /// Hue, Lightness and Saturation
     Hls,
 }
 
 impl PaletteInterpretation {
+    /// Creates a Rust [`PaletteInterpretation`] from  a C API [`GDALPaletteInterp`] value.
     fn from_c_int(palette_interpretation: GDALPaletteInterp::Type) -> Self {
         match palette_interpretation {
             GDALPaletteInterp::GPI_Gray => Self::Gray,
@@ -829,13 +844,25 @@ impl PaletteInterpretation {
             _ => unreachable!("GDAL has implemented a new type of `GDALPaletteInterp`"),
         }
     }
+
+    /// Returns the C API int value of this palette interpretation.
+    pub fn c_int(&self) -> GDALPaletteInterp::Type {
+        match self {
+            Self::Gray => GDALPaletteInterp::GPI_Gray,
+            Self::Rgba => GDALPaletteInterp::GPI_RGB,
+            Self::Cmyk => GDALPaletteInterp::GPI_CMYK,
+            Self::Hls => GDALPaletteInterp::GPI_HLS,
+        }
+    }
 }
 
+/// Grayscale [`ColorTable`] entry.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct GrayEntry {
     pub g: i16,
 }
 
+/// Red, green, blue, alpha [`ColorTable`] entry.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct RgbaEntry {
     pub r: i16,
@@ -844,6 +871,7 @@ pub struct RgbaEntry {
     pub a: i16,
 }
 
+/// Cyan, magenta, yellow, black [`ColorTable`] entry.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct CmykEntry {
     pub c: i16,
@@ -852,6 +880,7 @@ pub struct CmykEntry {
     pub k: i16,
 }
 
+/// Hue, lightness, saturation [`ColorTable`] entry.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct HlsEntry {
     pub h: i16,
@@ -859,7 +888,8 @@ pub struct HlsEntry {
     pub s: i16,
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+/// Options for defining [`ColorInterpretation::PaletteIndex`] entries in a [`ColorTable`].
+#[derive(Copy, Clone, PartialEq, Eq)]
 pub enum ColorEntry {
     Gray(GrayEntry),
     Rgba(RgbaEntry),
@@ -867,23 +897,213 @@ pub enum ColorEntry {
     Hls(HlsEntry),
 }
 
-/// Color table for raster bands that use the PaletteIndex color interpretation.
+impl ColorEntry {
+    /// Instantiate a greyscale color entry
+    pub fn grey(g: i16) -> Self {
+        Self::Gray(GrayEntry { g })
+    }
+
+    /// Instantiate an red, green, blue, alpha color entry
+    pub fn rgba(r: i16, g: i16, b: i16, a: i16) -> Self {
+        Self::Rgba(RgbaEntry { r, g, b, a })
+    }
+
+    /// Instantiate a cyan, magenta, yellow, black color entry
+    pub fn cmyk(c: i16, m: i16, y: i16, k: i16) -> Self {
+        Self::Cmyk(CmykEntry { c, m, y, k })
+    }
+
+    /// Instantiate a hue, lightness, saturation color entry
+    pub fn hls(h: i16, l: i16, s: i16) -> Self {
+        Self::Hls(HlsEntry { h, l, s })
+    }
+
+    /// Get the ['PaletteInterpretation'] describing `self`.
+    pub fn palette_interpretation(&self) -> PaletteInterpretation {
+        match self {
+            ColorEntry::Gray(_) => PaletteInterpretation::Gray,
+            ColorEntry::Rgba(_) => PaletteInterpretation::Rgba,
+            ColorEntry::Cmyk(_) => PaletteInterpretation::Cmyk,
+            ColorEntry::Hls(_) => PaletteInterpretation::Hls,
+        }
+    }
+
+    /// Create from a C [`GDALColorEntry`].
+    fn from(e: GDALColorEntry, interp: PaletteInterpretation) -> ColorEntry {
+        match interp {
+            PaletteInterpretation::Gray => ColorEntry::Gray(GrayEntry { g: e.c1 }),
+            PaletteInterpretation::Rgba => ColorEntry::Rgba(RgbaEntry {
+                r: e.c1,
+                g: e.c2,
+                b: e.c3,
+                a: e.c4,
+            }),
+            PaletteInterpretation::Cmyk => ColorEntry::Cmyk(CmykEntry {
+                c: e.c1,
+                m: e.c2,
+                y: e.c3,
+                k: e.c4,
+            }),
+            PaletteInterpretation::Hls => ColorEntry::Hls(HlsEntry {
+                h: e.c1,
+                l: e.c2,
+                s: e.c3,
+            }),
+        }
+    }
+}
+
+impl From<&ColorEntry> for GDALColorEntry {
+    fn from(e: &ColorEntry) -> Self {
+        match e {
+            ColorEntry::Gray(e) => GDALColorEntry {
+                c1: e.g,
+                c2: 0,
+                c3: 0,
+                c4: 0,
+            },
+            ColorEntry::Rgba(e) => GDALColorEntry {
+                c1: e.r,
+                c2: e.g,
+                c3: e.b,
+                c4: e.a,
+            },
+            ColorEntry::Cmyk(e) => GDALColorEntry {
+                c1: e.c,
+                c2: e.m,
+                c3: e.y,
+                c4: e.k,
+            },
+            ColorEntry::Hls(e) => GDALColorEntry {
+                c1: e.h,
+                c2: e.l,
+                c3: e.s,
+                c4: 0,
+            },
+        }
+    }
+}
+
+// For more compact debug output, skip enum wrapper.
+impl Debug for ColorEntry {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ColorEntry::Gray(e) => e.fmt(f),
+            ColorEntry::Rgba(e) => e.fmt(f),
+            ColorEntry::Cmyk(e) => e.fmt(f),
+            ColorEntry::Hls(e) => e.fmt(f),
+        }
+    }
+}
+
+/// Color table for raster bands that use [`ColorInterpretation::PaletteIndex`] color interpretation.
 ///
-/// This object carries the lifetime of the raster band that
-/// contains it. This is necessary to prevent the raster band
-/// from being dropped before the color table.
+///
+/// # Example
+///
+/// ```rust, no_run
+/// use gdal::{Dataset, Driver};
+/// use gdal::raster::{ColorEntry, ColorTable, PaletteInterpretation};
+/// # fn main() -> gdal::errors::Result<()> {
+///
+/// // Open source multinomial classification raster
+/// let ds = Dataset::open("fixtures/labels.tif")?;
+///
+/// // Create in-memory copy to mutate
+/// let mem_driver = Driver::get_by_name("MEM")?;
+/// let ds = ds.create_copy(&mem_driver, "<mem>", &[])?;
+/// let mut band = ds.rasterband(1)?;
+/// assert!(band.color_table().is_none());
+///
+/// // Create a new color table for 3 classes + no-data
+/// let mut ct = ColorTable::new(PaletteInterpretation::Rgba);
+/// ct.set_color_entry(0, &ColorEntry::rgba(255, 255, 0, 255));
+/// ct.set_color_entry(1, &ColorEntry::rgba(0, 255, 255, 255));
+/// ct.set_color_entry(2, &ColorEntry::rgba(255, 0, 255, 255));
+/// ct.set_color_entry(255, &ColorEntry::rgba(0, 0, 0, 0));
+/// band.set_color_table(&ct);
+///
+/// // Render a PNG
+/// let png_driver = Driver::get_by_name("PNG")?;
+/// ds.create_copy(&png_driver, "/tmp/labels.png", &[])?;
+///
+/// # Ok(())
+/// # }
+/// ```
 pub struct ColorTable<'a> {
     palette_interpretation: PaletteInterpretation,
     c_color_table: GDALColorTableH,
+    /// If `true`, Rust is responsible for deallocating color table pointed to by
+    /// `c_color_table`, which is the case when instantiated directly, as opposed to
+    /// when read via [`RasterBand::color_table`].
+    rust_owned: bool,
     phantom_raster_band: PhantomData<&'a RasterBand<'a>>,
 }
 
 impl<'a> ColorTable<'a> {
+    /// Instantiate a new color table with the given palette interpretation.
+    pub fn new(interp: PaletteInterpretation) -> Self {
+        let c_color_table = unsafe { GDALCreateColorTable(interp.c_int()) };
+        Self {
+            palette_interpretation: interp,
+            c_color_table,
+            rust_owned: true,
+            phantom_raster_band: PhantomData,
+        }
+    }
+
+    /// Constructs a color ramp from one color entry to another.
+    ///
+    /// `start_index` and `end_index` must be `0..=255`.
+    ///
+    /// Returns `None` if `start_color` and `end_color` do not have the same [`PaletteInterpretation`].
+    ///
+    /// # Example
+    ///
+    /// ```rust, no_run
+    /// use gdal::{Dataset, Driver};
+    /// use gdal::raster::{ColorEntry, ColorTable, PaletteInterpretation};
+    /// # fn main() -> gdal::errors::Result<()> {
+    /// // Create a 16 step blue to white color table.
+    /// let ct = ColorTable::color_ramp(
+    ///     0, &ColorEntry::rgba(0, 0, 255, 255),
+    ///     15, &ColorEntry::rgba(255, 255, 255, 255)
+    /// )?;
+    /// println!("{ct:?}");
+    /// # Ok(())
+    /// # }
+    pub fn color_ramp(
+        start_index: u8,
+        start_color: &ColorEntry,
+        end_index: u8,
+        end_color: &ColorEntry,
+    ) -> Result<ColorTable<'a>> {
+        if start_color.palette_interpretation() != end_color.palette_interpretation() {
+            Err(GdalError::BadArgument(
+                "start_color and end_color must have the same palette_interpretation".into(),
+            ))
+        } else {
+            let ct = ColorTable::new(start_color.palette_interpretation());
+            unsafe {
+                GDALCreateColorRamp(
+                    ct.c_color_table,
+                    start_index as c_int,
+                    &start_color.into(),
+                    end_index as c_int,
+                    &end_color.into(),
+                );
+            }
+            Ok(ct)
+        }
+    }
+
+    /// Wrap C color table
     fn from_c_color_table(c_color_table: GDALColorTableH) -> Self {
-        let interp_index = unsafe { gdal_sys::GDALGetPaletteInterpretation(c_color_table) };
+        let interp_index = unsafe { GDALGetPaletteInterpretation(c_color_table) };
         ColorTable {
             palette_interpretation: PaletteInterpretation::from_c_int(interp_index),
             c_color_table,
+            rust_owned: false,
             phantom_raster_band: PhantomData,
         }
     }
@@ -907,29 +1127,12 @@ impl<'a> ColorTable<'a> {
             }
             *c_color_entry
         };
-        match self.palette_interpretation {
-            PaletteInterpretation::Gray => Some(ColorEntry::Gray(GrayEntry { g: color_entry.c1 })),
-            PaletteInterpretation::Rgba => Some(ColorEntry::Rgba(RgbaEntry {
-                r: color_entry.c1,
-                g: color_entry.c2,
-                b: color_entry.c3,
-                a: color_entry.c4,
-            })),
-            PaletteInterpretation::Cmyk => Some(ColorEntry::Cmyk(CmykEntry {
-                c: color_entry.c1,
-                m: color_entry.c2,
-                y: color_entry.c3,
-                k: color_entry.c4,
-            })),
-            PaletteInterpretation::Hls => Some(ColorEntry::Hls(HlsEntry {
-                h: color_entry.c1,
-                l: color_entry.c2,
-                s: color_entry.c3,
-            })),
-        }
+        Some(ColorEntry::from(color_entry, self.palette_interpretation))
     }
 
     /// Get a color entry as RGB.
+    ///
+    /// Returns `None` if `palette_interpretation` is not `Rgba`.
     pub fn entry_as_rgb(&self, index: usize) -> Option<RgbaEntry> {
         let mut color_entry = GDALColorEntry {
             c1: 0,
@@ -949,5 +1152,42 @@ impl<'a> ColorTable<'a> {
             b: color_entry.c3,
             a: color_entry.c4,
         })
+    }
+
+    /// Set entry in the RasterBand color table.
+    ///
+    /// The `entry` variant type must match `palette_interpretation`.
+    ///
+    /// The table is grown as needed to hold the supplied index, filling in gaps with
+    /// the default [`GDALColorEntry`] value.
+    pub fn set_color_entry(&mut self, index: u16, entry: &ColorEntry) {
+        unsafe { GDALSetColorEntry(self.c_color_table, index as c_int, &entry.into()) }
+    }
+}
+
+impl Drop for ColorTable<'_> {
+    fn drop(&mut self) {
+        if self.rust_owned {
+            unsafe { GDALDestroyColorTable(self.c_color_table) }
+        }
+    }
+}
+
+impl Default for ColorTable<'_> {
+    fn default() -> Self {
+        Self::new(PaletteInterpretation::Rgba)
+    }
+}
+
+impl Debug for ColorTable<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let entries = (0..self.entry_count())
+            .filter_map(|i| self.entry(i))
+            .collect::<Vec<_>>();
+
+        f.debug_struct("ColorTable")
+            .field("palette_interpretation", &self.palette_interpretation)
+            .field("entries", &entries)
+            .finish()
     }
 }