+++ CHANGES.md
@@ -27,6 +27,11 @@
 
   - <https://github.com/georust/gdal/pull/311>
 
+- Added ability to set color table for bands with palette color interpretation. 
+  Added ability to create a color ramp (interpolated) color table.
+
+  - <https://github.com/georust/gdal/pull/314>
+
 ## 0.13
 
 - Add prebuild bindings for GDAL 3.5